requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
tavily-python